Starts to automate globals for go code generation so we can add a new global: chain
