+++ src/apps/blockScrape/handle_block_scrape.cpp
@@ -40,7 +40,7 @@ bool COptions::scrape_blocks(void) {
     // ...but we may make some adjustments to speed things up. When not running in docker mode,
     // we can do more blocks. In docker mode, we stick with the defaults otherwise, docker
     // may kill us for using too many resources.
-    if (!isDockerMode()) {
+    if (!isDockerMode() && chain == "mainnet") {
         if (cons.blazeStart < 450000) {
             // We can speed things up on the early chain...
             cons.blazeCnt = max(blknum_t(4000), cons.blazeCnt);
