+++ src/apps/blockScrape/options.cpp
@@ -205,6 +205,8 @@ void COptions::Init(void) {
     // clang-format on
     // END_CODE_INIT
 
+    chain = getGlobalConfig()->getConfigStr("settings", "chain", "mainnet");
+
     minArgs = 0;
 }
