+++ src/apps/blockScrape/options.h
@@ -38,6 +38,7 @@ class COptions : public COptionsBase {
     blknum_t latestBlockNum;
     CPinnedChunkArray pinList;
     CApiKey lic;
+    string_q chain;
 
     COptions(void);
     ~COptions(void);
