+++ src/apps/chifra/server/caller.go
@@ -150,61 +150,3 @@ func CallOne(w http.ResponseWriter, r *http.Request, tbCmd, extra, apiCmd string
 	}
 	fmt.Fprint(w, output)
 }
-
-// TODO: Once we're fully ported to Go, this can go away when CallOne goes away
-// TODO: camelCase vs. snake_case - issue #1974
-func convertToCommandLine(in string) string {
-	switch in {
-	case "addrChanCnt":
-		return "addr_chan_cnt"
-	case "bigRange":
-		return "big_range"
-	case "blockChanCnt":
-		return "block_chan_cnt"
-	case "blockCnt":
-		return "block_cnt"
-	case "byAcct":
-		return "by_acct"
-	case "byDate":
-		return "by_date"
-	case "cacheTraces":
-		return "cache_traces"
-	case "firstBlock":
-		return "first_block"
-	case "firstRecord":
-		return "first_record"
-	case "getConfig":
-		return "get_config"
-	case "initAll":
-		return "init_all"
-	case "lastBlock":
-		return "last_block"
-	case "listCount":
-		return "list_count"
-	case "matchCase":
-		return "match_case"
-	case "maxRecords":
-		return "max_records"
-	case "maxTraces":
-		return "max_traces"
-	case "noZero":
-		return "no_zero"
-	case "proxyFor":
-		return "proxy_for"
-	case "setConfig":
-		return "set_config"
-	case "skipDdos":
-		return "skip_ddos"
-	case "summarizeBy":
-		return "summarize_by"
-	case "toCustom":
-		return "to_custom"
-	case "logLevel":
-		return "log_level"
-	case "noHeader":
-		return "no_header"
-	case "toFile":
-		return "to_file"
-	}
-	return in
-}
